Got it working with 2 digits!
So the base for the bignums is now 100, each bignum digit is two decimal
digits, so it should work faster, but I haven't actually measured.

The main reason this seemed to be breaking before is not because I was
computing it wrong or breaking up the strings incorrectly, but because
each bignum digit wasn't printed with leading zeros, so putting them all
together ended up wrong. Easy enough to fix with a map and sprintf.